% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rxylib.R
\name{run_rxylib}
\alias{run_rxylib}
\title{Run Luminescence shiny apps}
\usage{
run_rxylib(...)
}
\arguments{
\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\description{
A wrapper for \code{\link{runApp}} to start interactive shiny apps for the R package Luminescence.
}
\details{
The \code{app_RLum()} function is just a wrapper for \code{\link{runApp}}.
Via the \code{...} argument further arguments can be directly passed to \code{\link{runApp}}.
See \code{?shiny::runApp} for further details on valid arguments.
}
\examples{

\dontrun{

run_rxylib()

}

}
\seealso{
\code{\link{runApp}}
}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
